refactor(actor-adaptor-std): clean up std file names
Keep Std* type names intact while removing redundant std_ file names under the std adapter surface.

Move StdTickDriver into tick_driver.rs so the module layout stays compliant with workspace wiring and lint rules.
