+++ modules/actor-adaptor-std/src/std.rs
@@ -1,12 +1,12 @@
 /// Actor-specific standard-library bindings.
 pub mod actor;
+mod blocker;
 /// Dispatch bindings for the standard toolbox.
 pub mod dispatch;
 /// Event bindings for the standard toolbox.
 pub mod event;
 /// Pattern bindings for the standard toolbox.
 pub mod pattern;
-mod std_blocker;
 /// Test-support helpers for actor systems (test-support feature only).
 #[cfg(feature = "test-support")]
 pub mod system;
@@ -15,4 +15,4 @@ pub mod tick_driver;
 /// Time bindings for the standard toolbox.
 pub mod time;
 
-pub use std_blocker::StdBlocker;
+pub use blocker::StdBlocker;
