+++ modules/actor-adaptor-std/src/std/system.rs
@@ -1,7 +1,7 @@
 //! System-level test helpers (gated by `test-support` feature at parent module).
 
+mod actor_system_config;
 mod empty_system;
-mod std_actor_system_config;
 
+pub use actor_system_config::std_actor_system_config;
 pub use empty_system::{new_empty_actor_system, new_empty_actor_system_with};
-pub use std_actor_system_config::std_actor_system_config;
