+++ modules/actor-adaptor-std/src/std/time.rs
@@ -1,7 +1,7 @@
 //! Time bindings for the standard toolbox.
 
-mod std_clock;
-mod std_mailbox_clock;
+mod clock;
+mod monotonic_mailbox_clock;
 
-pub use std_clock::StdClock;
-pub use std_mailbox_clock::std_monotonic_mailbox_clock;
+pub use clock::StdClock;
+pub use monotonic_mailbox_clock::std_monotonic_mailbox_clock;
